+++ CompService.Core/Models/SparePart.cs
@@ -7,6 +7,6 @@ public class SparePart
     public string Article { get; set; } = null!;
     public SparePartCategory Category { get; set; } = null!;
     public int Count { get; set; }
-    public decimal PurchasePrice { get; set; }
-    public decimal RetailPrice { get; set; }
+    public double PurchasePrice { get; set; }
+    public double RetailPrice { get; set; }
 }
